Learned about tuple and array. To create a tuple use () and to access the element in tuple use indexing method like if a tuple has t=(2,3,4) the t.0 will give answer 2. and println! macro's {} will expects either variable names or expressions, but not field accesses like tup.0. so use println!("{}",tup.0);
For array use [] to initialize and let arr[3; 5] will create an array of size 5 with all initialized with 3.

Indexing with u32: In Rust, array indices must be of type usize. Using u32 will result in a compilation error.

So the index value should be in a type of usize not others.